python/requirements.txt
pyserial==3.5
pyqtgraph==0.13.3
PyQt5==5.15.9
